src/weather-agent-python/pyproject.toml
[project]
name = "weather-agent-python"
version = "0.1.0"
description = "Weather intelligence agent for AlpineAI ski resort"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "opentelemetry-api>=1.33.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.0",
    "opentelemetry-instrumentation-fastapi>=0.54b0",
    "opentelemetry-sdk>=1.33.0",
    "grpcio>=1.50.0",
    "agent-framework",
    "agent-framework-azure",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
prerelease = "allow"

[project.scripts]
start = "weather_agent_python.main:main"
